+++ Mapper/Mapper.swift
@@ -104,6 +104,24 @@ public struct Mapper {
         return try? self.from(field)
     }
 
+    /**
+     Get an optional value from the given keys and source data. This returns the first non-nil value produced
+     in order based on the array of fields
+
+     - parameter fields: The array of fields to check from the source data.
+
+     - returns: The first non-nil value to be produced from the array of fields, or nil if none exist
+    */
+    public func optionalFrom<T: RawRepresentable>(fields: [String]) -> T? {
+        for field in fields {
+            if let value: T = try? self.from(field) {
+                return value
+            }
+        }
+
+        return nil
+    }
+
     // MARK: - T: Mappable
 
     /**
@@ -179,6 +197,24 @@ public struct Mapper {
         return try? self.from(field)
     }
 
+    /**
+     Get an optional value from the given keys and source data. This returns the first non-nil value produced
+     in order based on the array of fields
+
+     - parameter fields: The array of fields to check from the source data.
+
+     - returns: The first non-nil value to be produced from the array of fields, or nil if none exist
+    */
+    public func optionalFrom<T: Mappable>(fields: [String]) -> T? {
+        for field in fields {
+            if let value: T = try? self.from(field) {
+                return value
+            }
+        }
+
+        return nil
+    }
+
     // MARK: - T: Convertible
 
     /**
@@ -243,6 +279,24 @@ public struct Mapper {
      */
     public func optionalFrom<T: Convertible where T == T.ConvertedType>(field: String) -> [T]? {
         return try? self.from(field)
+    }
+
+    /**
+     Get an optional value from the given keys and source data. This returns the first non-nil value produced
+     in order based on the array of fields
+
+     - parameter fields: The array of fields to check from the source data.
+
+     - returns: The first non-nil value to be produced from the array of fields, or nil if none exist
+    */
+    public func optionalFrom<T: Convertible where T == T.ConvertedType>(fields: [String]) -> T? {
+        for field in fields {
+            if let value: T = try? self.from(field) {
+                return value
+            }
+        }
+
+        return nil
     }
 
     // MARK: - Custom Transformation