+++ MapperTests/ConvertibleValueTests.swift
@@ -97,4 +97,16 @@ final class ConvertibleValueTests: XCTestCase {
         let test = try! Test(map: Mapper(JSON: ["urls": "not an array"]))
         XCTAssertNil(test.URLs)
     }
+
+    func testConvertibleArrayOfKeys() {
+        struct Test: Mappable {
+            let URL: NSURL?
+            init(map: Mapper) throws {
+                self.URL = map.optionalFrom(["a", "b"])
+            }
+        }
+
+        let test = try! Test(map: Mapper(JSON: ["a": "##", "b": "example.com"]))
+        XCTAssertTrue(test.URL?.absoluteString == "example.com")
+    }
 }