+++ MapperTests/MappableValueTests.swift
@@ -131,4 +131,23 @@ final class MappableValueTests: XCTestCase {
         let test = try! Test(map: Mapper(JSON: ["nests": "not an array"]))
         XCTAssertNil(test.nests)
     }
+
+    func testMappableArrayOfKeys() {
+        struct Test: Mappable {
+            let nest: Nested?
+            init(map: Mapper) throws {
+                self.nest = map.optionalFrom(["a", "b"])
+            }
+        }
+
+        struct Nested: Mappable {
+            let string: String
+            init(map: Mapper) throws {
+                try self.string = map.from("string")
+            }
+        }
+
+        let test = try! Test(map: Mapper(JSON: ["a": ["foo": "bar"], "b": ["string": "hi"]]))
+        XCTAssertTrue(test.nest?.string == "hi")
+    }
 }