+++ CHANGELOG.md
@@ -10,7 +10,10 @@
 
 ## Bug Fixes
 
-- None
+- Fixed issue where `optionalFrom` with an array of keys was not
+  implemented for `Convertible` or other types.
+  [Keith Smiley](https://github.com/keith)
+  [#10](https://github.com/lyft/mapper/pull/10)
 
 # 1.0.0